+++ autd3-gain-holo/src/linear_synthesis/gspat.rs
@@ -52,7 +52,7 @@ impl<D: Directivity, B: LinAlgBackend<D>> GSPAT<D, B> {
         let mut q = self.backend.alloc_zeros_cv(n)?;
 
         let amps = self.backend.from_slice_cv(unsafe {
-            std::slice::from_raw_parts(self.amps.as_ptr() as *const f64, self.amps.len())
+            std::slice::from_raw_parts(self.amps.as_ptr() as *const f32, self.amps.len())
         })?;
 
         let b = self.backend.gen_back_prop(n, m, &g)?;
